target/ppc: Move SPR generation to separate file
This move is required to enable building without TCG.
All the logic related to registering SPRs specific to
some architectures or machines has been hidden in this
new file.

The idea of this final patch is to hide all SPR generation from
translate_init, but in an effort to simplify the RFC the 4
functions for not_implemented SPRs were created. They'll be
substituted by gen_spr_<machine>_misc in reusable ways for the
final patch.

another issue we ran into was vscr_init using static functions
means it has to be static, so we had to remove them from
gen_spr_74xx and gen_spr_book3s_altivec, and have them in
the init_procs instead.

Finally, SPR_NOACCESS had to be defined in internal.h, as it
is used by spr_common, translate_init and translate. If there
is a better solution, I'll be happy to implement it.

As for the redundant code complaint this patch will get, it has only
been moved, so I don't know if I can remove that code

Signed-off-by: Bruno Larsen (billionai) <bruno.larsen@eldorado.org.br>